[216] Extract shared Synthefy errors (#431)
* Extract shared Synthefy errors

* Include shared errors in artifact manifest
