+++ libs/synthefy/SOURCE_SNAPSHOT.json
@@ -165,6 +165,15 @@
           "changes": [
             "replace evaluated PEP 604 annotations with Optional for Python 3.9"
           ]
+        },
+        {
+          "input_blob": "e108b75e10942c51a8af17cf8e4e34b930d838e3",
+          "result_blob": "0765ce2a1c1df3d168ea09793ceb2424ebfab921",
+          "phase": "shared_error_ownership",
+          "decision_record": "../../docs/architecture/0001-consolidate-synthefy-source-tree-ci-gates.json",
+          "changes": [
+            "re-export shared SDK errors and status mapping from their dependency-light owner"
+          ]
         }
       ],
       "src/synthefy/nori_client.py": [
@@ -197,6 +206,15 @@
           "changes": [
             "remove auto from the supported execution mode type and resolver",
             "reject the retired value before credentials, model selection, or transport setup"
+          ]
+        },
+        {
+          "input_blob": "4834f552c280ac6dca7f2b03e4c6665c9cf68066",
+          "result_blob": "73db4cb6436f7a83d69cca0204a9848cf93c01c7",
+          "phase": "shared_error_ownership",
+          "decision_record": "../../docs/architecture/0001-consolidate-synthefy-source-tree-ci-gates.json",
+          "changes": [
+            "import shared exceptions and HTTP status mapping without loading the legacy ForecastV2 client"
           ]
         }
       ],
@@ -253,6 +271,19 @@
       ]
     },
     "source_relocations": {
+      "src/synthefy/errors.py": {
+        "source_project_path": "src/synthefy/api_client.py",
+        "source_commit": "6438cfc0ca5a685842a5fe86b4fe9cfa8b10d07d",
+        "source_blob": "e108b75e10942c51a8af17cf8e4e34b930d838e3",
+        "result_blob": "22f24f027a8e312eb19bb3a8c67635e9f887f599",
+        "sha256": "ada35c2bfe8b707c1bad46b864363021be6fd80b12efbab6864d9370b39a15f4",
+        "phase": "shared_error_ownership",
+        "decision_record": "../../docs/architecture/0001-consolidate-synthefy-source-tree-ci-gates.json",
+        "changes": [
+          "extract the existing exception hierarchy and HTTP status mapping from the legacy ForecastV2 client",
+          "preserve every class name, constructor, response field, and status mapping"
+        ]
+      },
       "tests/test_featurize.py": {
         "source_project_path": "../../tests/test_featurize.py",
         "source_commit": "85c083c8f3fca791c9959ca5fd0d4a83eeb2492a",
@@ -455,6 +486,7 @@
         "src/synthefy/__init__.py",
         "src/synthefy/api_client.py",
         "src/synthefy/data_models.py",
+        "src/synthefy/errors.py",
         "src/synthefy/featurize.py",
         "src/synthefy/nori_client.py",
         "src/synthefy/nori_data_models.py",
@@ -473,6 +505,7 @@
         "synthefy/__init__.py",
         "synthefy/api_client.py",
         "synthefy/data_models.py",
+        "synthefy/errors.py",
         "synthefy/featurize.py",
         "synthefy/nori_client.py",
         "synthefy/nori_data_models.py",
